LPS-56341 Moves productivity-center-web to product-menu-web
