+++ modules/apps/build.xml
@@ -37,7 +37,6 @@
 		<dirset dir="portlet-css" includes="*" />
 		<dirset dir="portlet-display-template" includes="*" />
 		<dirset dir="product-menu" includes="*" />
-		<dirset dir="productivity-center" includes="*" />
 		<dirset dir="quick-note" includes="*" />
 		<dirset dir="roles" includes="*" />
 		<dirset dir="rss" includes="*" />
