+++ modules/apps/product-menu/product-menu-web/bnd.bnd
@@ -1,5 +1,5 @@
-Bundle-Name: Liferay Productivity Center Web
-Bundle-SymbolicName: com.liferay.productivity.center.web
+Bundle-Name: Liferay Product Menu Web
+Bundle-SymbolicName: com.liferay.product.menu.web
 Bundle-Version: 1.0.0
 Import-Package:\
 	com.liferay.application.list.taglib.servlet.taglib.ui,\
@@ -9,4 +9,4 @@ Include-Resource:\
 	META-INF/resources=src/META-INF/resources
 Require-Capability:\
 	osgi.extender;osgi.extender="jsp.taglib";uri="http://liferay.com/tld/application-list/ui";version:Version="1.0.0"
-Web-ContextPath: /productivity-center-web
+Web-ContextPath: /product-menu-web
