+++ modules/apps/product-menu/product-menu-web/src/META-INF/resources/layout/view/user_personal_panel.jsp
