+++ modules/apps/product-menu/product-menu-web/src/META-INF/resources/portlet/init.jsp
