+++ modules/apps/product-menu/product-menu-web/src/META-INF/resources/portlet/view.jsp
