+++ modules/apps/product-menu/product-menu-web/src/com/liferay/product/menu/web/constants/ProductivityCenterPortletKeys.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.productivity.center.constants;
+package com.liferay.product.menu.web.constants;
 
 /**
  * @author Adolfo Pérez
