+++ modules/apps/product-menu/product-menu-web/src/com/liferay/product/menu/web/portlet/ProductivityCenterPortlet.java
@@ -12,15 +12,12 @@
  * details.
  */
 
-package com.liferay.productivity.center.layout;
+package com.liferay.product.menu.web.portlet;
 
-import com.liferay.portal.model.Layout;
-import com.liferay.portal.model.LayoutConstants;
-import com.liferay.portal.model.LayoutTypeAccessPolicy;
-import com.liferay.portal.model.Portlet;
-import com.liferay.portal.model.impl.DefaultLayoutTypeAccessPolicyImpl;
+import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
+import com.liferay.product.menu.web.constants.ProductivityCenterPortletKeys;
 
-import javax.servlet.http.HttpServletRequest;
+import javax.portlet.Portlet;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -29,15 +26,14 @@ import org.osgi.service.component.annotations.Component;
  */
 @Component(
 	immediate = true,
-	property = {"layout.type=" + LayoutConstants.TYPE_USER_PERSONAL_PANEL},
-	service = LayoutTypeAccessPolicy.class
+	property = {
+		"com.liferay.portlet.display-category=category.hidden",
+		"com.liferay.portlet.instanceable=false",
+		"javax.portlet.init-param.view-template=/portlet/view.jsp",
+		"javax.portlet.name=" + ProductivityCenterPortletKeys.PRODUCTIVITY_CENTER,
+		"javax.portlet.supports.mime-type=text/html"
+	},
+	service = Portlet.class
 )
-public class UserPersonalPanelLayoutTypeAccessPolicy
-	extends DefaultLayoutTypeAccessPolicyImpl {
-
-	@Override
-	public void checkAccessAllowedToPortlet(
-		HttpServletRequest request, Layout layout, Portlet portlet) {
-	}
-
+public class ProductivityCenterPortlet extends MVCPortlet {
 }
