--- modules/apps/productivity-center/build.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE project>
-
-<project>
-	<import file="../../../tools/sdk/build-common-plugins.xml" />
-</project>
