--- modules/apps/productivity-center/productivity-center-web/snapshot.properties
@@ -1,4 +1,0 @@
-
-module.snapshot.git.hash=bb7af5ab7a7743d1ce8d6d575a66b5f3b4462339
-module.snapshot.url=http://cdn.repository.liferay.com/nexus/content/repositories/liferay-snapshots-ce/com/liferay/com.liferay.productivity.center.web/1.0.0-SNAPSHOT/com.liferay.productivity.center.web-1.0.0-20150625.044008-8.jar
-
